directory structures updates and loop-run preparations
