+++ log_examples/statistic_output.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!-- generated on 2024-01-08 04:49:52 by Eclipse SUMO sumo Version 1.18.0
+<configuration xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="http://sumo.dlr.de/xsd/sumoConfiguration.xsd">
+
+    <input>
+        <net-file value="/home/pavel/dev/diplom/tssproblem/medium/sumo/../net/osm.net.xml"/>
+        <route-files value="/home/pavel/dev/diplom/tssproblem/medium/duaIterate/iterations/299/trips.odtrips_299.rou.xml"/>
+        <additional-files value="/home/pavel/dev/diplom/tssproblem/medium/sumo/../polygons/osm.poly.xml"/>
+    </input>
+
+    <output>
+        <statistic-output value="statistic_output.xml"/>
+    </output>
+
+    <processing>
+        <ignore-route-errors value="true"/>
+    </processing>
+
+    <routing>
+        <device.rerouting.adaptation-steps value="18"/>
+        <device.rerouting.adaptation-interval value="10"/>
+    </routing>
+
+    <report>
+        <verbose value="true"/>
+        <duration-log.statistics value="true"/>
+        <no-step-log value="true"/>
+    </report>
+
+    <gui_only>
+        <gui-settings-file value="/home/pavel/dev/diplom/tssproblem/medium/sumo/osm.view.xml"/>
+    </gui_only>
+
+</configuration>
+-->
+
+<statistics xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="http://sumo.dlr.de/xsd/statistic_file.xsd">
+    <performance clockBegin="1704678592.67" clockEnd="1704678596.29" clockDuration="3.62" traciDuration="2.32" realTimeFactor="2045.84" vehicleUpdatesPerSecond="238262.91" personUpdatesPerSecond="0.00" begin="0.00" end="7408.00" duration="7408.00"/>
+    <vehicles loaded="3000" inserted="3000" running="0" waiting="0"/>
+    <teleports total="2" jam="0" yield="1" wrongLane="1"/>
+    <safety collisions="0" emergencyStops="0" emergencyBraking="0"/>
+    <persons loaded="0" running="0" jammed="0"/>
+    <personTeleports total="0" abortWait="0" wrongDest="0"/>
+    <vehicleTripStatistics count="3000" routeLength="945.78" speed="5.24" duration="287.58" waitingTime="152.55" timeLoss="229.56" departDelay="606.61" departDelayWaiting="-1.00" totalTravelTime="862750.00" totalDepartDelay="1819816.99"/>
+    <pedestrianStatistics number="0" routeLength="0.00" duration="0.00" timeLoss="0.00"/>
+    <rideStatistics number="0"/>
+    <transportStatistics number="0"/>
+</statistics>
