--- report.log
@@ -1,42 +1,0 @@
-2024-01-10 05:51:32,293 - pyswarms.single.global_best - INFO - Optimize for 5 iters with {'c1': 0.5, 'c2': 0.3, 'w': 0.9}
-2024-01-10 05:53:43,391 - pyswarms.single.global_best - INFO - Optimization finished | best cost: 65.8, best pos: [41.52791525  3.00585485 38.8525078   2.96225063 37.28470065  2.97065316
- 51.37852535  2.96705193 48.35500024  3.01538293 48.45790298  3.04184086
- 56.11960397  3.00890885 52.97663861  2.99984434 32.80378758  2.96074671
- 47.36270649  3.03260645 57.20669211  3.00541966 54.0286712  45.07961343
-  2.97122191 39.68311056 35.47785018  2.98350786 42.44500589 53.39733876
-  2.99922488 43.96364618 36.74839871  3.00266326 35.70117106 50.73242679
-  3.03404937 44.22470757]
-2024-01-10 19:25:38,144 - pyswarms.single.global_best - INFO - Optimize for 5 iters with {'c1': 0.5, 'c2': 0.3, 'w': 0.9}
-2024-01-10 19:27:40,882 - pyswarms.single.global_best - INFO - Optimization finished | best cost: 49.32, best pos: [53.1701339   3.04956268 59.19911957  2.99751256 31.61136116  3.03130281
- 55.45049922  2.97563381 57.43097119  2.95967253 39.36992391  3.01079157
- 35.54633043  2.9578615  58.7245076   3.01195348 35.89283568  2.95429788
- 47.49336615  3.01629526 58.93870595  2.97129117 40.40530143 57.76391262
-  3.03882109 44.35282082 41.36759524  3.02048796 40.71811633 52.4931678
-  2.98909975 36.84640086 52.5590624   2.98780366 55.91890514 53.22588587
-  3.03411158 36.08048764]
-2024-01-11 01:34:21,521 - pyswarms.single.global_best - INFO - Optimize for 3 iters with {'c1': 0.5, 'c2': 0.3, 'w': 0.9}
-2024-01-11 01:34:55,113 - pyswarms.single.global_best - INFO - Optimization finished | best cost: 68.22, best pos: [46.40320876  3.00061044 46.10503762  2.9850611  52.11375385  2.99779922
- 37.64215029  2.95622899 47.33723509  2.95778155 46.04225941  2.97646635
- 37.63441628  3.01569608 50.71294499  3.00819584 45.91960673  3.04019418
- 46.82902535  3.01293944 38.90218081  2.95219904 37.89018104 45.89704615
-  2.99336833 42.99518928 49.20329903  2.95922305 59.54265005 45.143732
-  2.98115508 33.60486081 44.36826142  3.01222201 37.49091382 48.8790053
-  3.01753976 46.39088554]
-2024-01-11 06:25:56,134 - pyswarms.single.global_best - INFO - Optimize for 3 iters with {'c1': 0.5, 'c2': 0.3, 'w': 0.9}
-2024-01-11 06:26:09,719 - pyswarms.single.global_best - INFO - Optimize for 3 iters with {'c1': 0.5, 'c2': 0.3, 'w': 0.9}
-2024-01-11 06:26:39,717 - pyswarms.single.global_best - INFO - Optimization finished | best cost: 60.39, best pos: [30.44892842  2.95595725 57.44499957  2.97072673 45.72099661  2.98679112
- 45.28437465  2.96378728 40.72480744  3.03556565 58.42300657  2.97979293
- 39.06762433  2.95612047 55.00673339  2.95704586 52.81276972  3.00775075
- 46.17861323  3.00004489 34.9007672   3.04559854 30.40096883 50.43818411
-  3.01922217 35.15415459 56.06806736  3.04258726 39.34922843 32.65710556
-  2.99849754 48.90616407 52.07809641  2.95285203 39.5180029  48.08114434
-  3.01498298 36.50001788]
-2024-01-11 06:26:57,456 - pyswarms.single.global_best - INFO - Optimize for 3 iters with {'c1': 0.5, 'c2': 0.3, 'w': 0.9}
-2024-01-11 06:28:08,734 - pyswarms.single.global_best - INFO - Optimization finished | best cost: 55.14, best pos: [59.51015415  2.98766297 35.60654053  2.96472565 43.53948905  2.95486989
- 39.40087348  3.02158598 36.69924264  3.00783984 49.02223831  3.01175356
- 56.43405428  3.03399216 34.05729792  2.95607617 34.08885981  2.96415272
- 58.23936731  2.97697373 44.82414247  2.97402168 31.79006032 57.78641477
-  3.02408802 36.28849226 34.31407099  3.01168531 53.20329004 57.25668048
-  2.99242555 30.06405137 38.49878923  2.97888246 31.13798371 49.42752203
-  2.95891907 44.0425396 ]
-2024-01-11 06:28:16,276 - pyswarms.single.global_best - INFO - Optimize for 3 iters with {'c1': 0.5, 'c2': 0.3, 'w': 0.9}
