--- test/duaIterate/clean.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-sudo rm -rf iterations
