--- test/od2trips/od2trips_run.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-od2trips -c od2trips.config.xml -n from_to.taz.xml -o trips.odtrips.xml
